school-management/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
dateparser==1.1.8
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-compat==1.0.15
django-grappelli==3.0.7
django-hijack==3.4.1
django-hijack-admin==2.1.10
django-ical==1.9.2
django-nose==1.4.7
django-recurrence==1.11.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
icalendar==5.0.8
idna==3.4
kombu==5.3.2
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
multidict==6.0.4
nose==1.3.7
nose-exclude==0.5.0
nose-timer==1.0.1
O365==2.0.28
oauthlib==3.2.2
pdfkit==1.0.0
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
py==1.11.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
redis==5.0.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
sentry-sdk==1.31.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.5
sqlparse==0.4.4
stringcase==1.2.0
twilio==8.8.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
unicodecsv==0.14.1
urllib3==2.0.5
vine==5.0.0
vobject==0.9.6.1
wcwidth==0.2.6
yarl==1.9.2
